@@ -1,9 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Observable } from "rxjs/Observable";
+import { Observable,of } from "rxjs";
 import { Actor } from '../common/actor';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/of';
+import { catchError, map, tap } from 'rxjs/operators';
 
 @Injectable()
 export class ActorsService {
@@ -13,21 +12,21 @@ export class ActorsService {
 
   public findActorById(id: number) :Observable<Actor> {
       if(!id && id !== 0) {
-          return Observable.of(null);
+          return of(null);
       }
-      return this.http.get('/rest/actor/id/'+id, this._reqOptionsArgs).catch(error => {
+      return this.http.get<Actor>('/rest/actor/id/'+id, this._reqOptionsArgs).pipe(catchError(error => {
           console.error( JSON.stringify( error ) );
           return Observable.throw( error );
-          });
+          }));
   }
   
   public findActorByName(name: string) :Observable<Actor[]> {
       if(!name) {
-          return Observable.of([]);
+          return of([]);
       }
-      return this.http.get('/rest/actor/'+name, this._reqOptionsArgs).catch(error => {
+      return this.http.get<Actor[]>('/rest/actor/'+name, this._reqOptionsArgs).pipe(catchError(error => {
           console.error( JSON.stringify( error ) );
           return Observable.throw( error );
-          });
+          }));
   }
 }
