@@ -1,9 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Observable } from "rxjs/Observable";
+import { Observable,of } from "rxjs";
 import { Actor } from '../common/actor';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/of';
+import { catchError, map, tap } from 'rxjs/operators';
 import { Movie } from '../common/movie';
 import { Genere } from '../common/genere';
 
@@ -14,60 +13,60 @@ export class MoviesService {
     constructor(private http: HttpClient) { }
     
     public findMoviesByGenereId(id: number) : Observable<Movie[]> {
-        return this.http.get('/rest/movie/genere/id/'+id, this._reqOptionsArgs).catch(error => {
+        return this.http.get<Movie[]>('/rest/movie/genere/id/'+id, this._reqOptionsArgs).pipe(catchError(error => {
             console.error( JSON.stringify( error ) );
             return Observable.throw( error );
-        });
+        }));
     }
     
     public findMovieById(id: number) :Observable<Movie> {
         if(!id && id !== 0) {
-            return Observable.of(null);
+            return of(null);
         }
-        return this.http.get('/rest/movie/id/'+id, this._reqOptionsArgs).catch(error => {
+        return this.http.get<Movie>('/rest/movie/id/'+id, this._reqOptionsArgs).pipe(catchError(error => {
             console.error( JSON.stringify( error ) );
             return Observable.throw( error );
-            });
+            }));
     }
 
     public deleteMovieById(id: number) :Observable<boolean> {
         if(!id && id !== 0) {
-            return Observable.of(false);
+            return of(false);
         }
-        return this.http.delete('/rest/movie/id/'+id, this._reqOptionsArgs).catch(error => {
+        return this.http.delete<boolean>('/rest/movie/id/'+id, this._reqOptionsArgs).pipe(catchError(error => {
             console.error( JSON.stringify( error ) );
             return Observable.throw( error );
-        });
+        }));
     }
     
     public findMovieByTitle(title: string) :Observable<Movie[]> {
         if(!title) {
-            return Observable.of([]);
+            return of([]);
         }
-        return this.http.get('/rest/movie/'+title, this._reqOptionsArgs).catch(error => {
+        return this.http.get<Movie[]>('/rest/movie/'+title, this._reqOptionsArgs).pipe(catchError(error => {
             console.error( JSON.stringify( error ) );
             return Observable.throw( error );
-            });
+            }));
     }
     
     public importMoveByTitle(title: string) :Observable<Movie[]> {
         if(!title) {
-            return Observable.of([]);
+            return of([]);
         }
-        return this.http.get('/rest/movie/import/'+title, this._reqOptionsArgs).catch(error => {
+        return this.http.get<Movie[]>('/rest/movie/import/'+title, this._reqOptionsArgs).pipe(catchError(error => {
             console.error( JSON.stringify( error ) );
             return Observable.throw( error );
-            });
+            }));
     }
     
     public importMoveByTitleAndId(title: string, id: number) :Observable<boolean> {
         if(!title) {
             console.log("title: "+title+" id: "+id);
-            return Observable.of(false);
+            return of(false);
         }
-        return this.http.get('/rest/movie/import/'+title+'/number/'+id, this._reqOptionsArgs).catch(error => {
+        return this.http.get<boolean>('/rest/movie/import/'+title+'/number/'+id, this._reqOptionsArgs).pipe(catchError(error => {
             console.error( JSON.stringify( error ) );
             return Observable.throw( error );
-            });
+            }));
     }
 }
