@@ -1,10 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Observable } from "rxjs/Observable";
+import { Observable,of } from "rxjs";
 import { Actor } from '../common/actor';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/observable/throw';
+import { catchError, map, tap } from 'rxjs/operators';
 import { User } from '../common/user';
 import { Genere } from '../common/genere';
 
@@ -15,37 +13,37 @@ export class UsersService {
     
   constructor(private http: HttpClient) { }
 
-  public login(login: string, password: string) :Observable<boolean | any> {
+  public login(login: string, password: string) :Observable<boolean> {
       if(!login || !password) {
-          return Observable.of(false);
+          return of(false);
       }
       let u = new User();
       u.username = login;
       u.password = password;
-      return this.http.post('/rest/user/login', u, this._reqOptionsArgs).catch(error => {
+      return this.http.post<boolean>('/rest/user/login', u, this._reqOptionsArgs).pipe(catchError(error => {
           console.error( JSON.stringify( error ) );
-          return Observable.of(false);         
-          });
+          return of(false);         
+          }));
   }
   
-  public signin(login: string, password: string, movieDbKey: string) :Observable<boolean | any> {
+  public signin(login: string, password: string, movieDbKey: string) :Observable<boolean> {
       if(!login || !password || !movieDbKey) {
-          return Observable.of(false);
+          return of(false);
       }
       let u = new User();
       u.username = login;
       u.password = password;
       u.moviedbkey = movieDbKey;
-      return this.http.post('/rest/user/signin', u, this._reqOptionsArgs).catch(error => {
+      return this.http.post<boolean>('/rest/user/signin', u, this._reqOptionsArgs).pipe(catchError(error => {
           console.error( JSON.stringify( error ) );
-          return Observable.of(false);          
-          });
+          return of(false);          
+          }));
   }
   
   public allGeneres() :Observable<Genere[]> {
-      return this.http.get('/rest/user/genere', this._reqOptionsArgs).catch(error => {
+      return this.http.get<Genere[]>('/rest/user/genere', this._reqOptionsArgs).pipe(catchError(error => {
           console.error( JSON.stringify( error ) );
           return Observable.throw( error );
-      });
+      }));
   }
 }
